Login admin
inscription mande

+++ src/main/java/com/bibliotheque/repository/AdherentRepository.java
@@ -3,5 +3,4 @@ package com.bibliotheque.repository;
 import com.bibliotheque.entity.Adherent;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AdherentRepository extends JpaRepository<Adherent, Integer> {
-}
+public interface AdherentRepository extends JpaRepository<Adherent, Long> { }
